Machine_Learning_Module/sentimentAnalysis/requirement.txt
google-cloud-storage
google-cloud-language
flask_cors